Feat: entity, repository 패키지 생성 및 파일 구조 변경, Repository 코드 작성,
