+++ src/main/java/com/odos/odos_server/diary/repository/DiaryImageRepository.java
@@ -1,0 +1,9 @@
+package com.odos.odos_server.diary.repository;
+
+import com.odos.odos_server.diary.entity.DiaryImage;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface DiaryImageRepository extends JpaRepository<DiaryImage, Long> {
+}
