+++ src/main/java/com/odos/odos_server/member/repository/MemberRepository.java
@@ -1,0 +1,9 @@
+package com.odos.odos_server.member.repository;
+
+import com.odos.odos_server.member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MemberRepository extends JpaRepository<Member, Long> {
+}
